# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Generate Token for Authentication",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT Token",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "output_schema": {
                                            "$ref": "#/definitions/dto.AuthResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    },
                    "401": {
                        "description": "Authentication Failed",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh JWT Token",
                "responses": {
                    "200": {
                        "description": "New JWT Token",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "output_schema": {
                                            "$ref": "#/definitions/dto.AuthResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Authentication Failed",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User Registration Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration Success",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    },
                    "409": {
                        "description": "User already exists",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/certification": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "certification"
                ],
                "summary": "Insert Certification",
                "parameters": [
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InsertCertificationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certification Details",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/certification/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "certification"
                ],
                "summary": "Update Certification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCertificationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certification resp",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/education": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "education"
                ],
                "summary": "Insert Education",
                "parameters": [
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InsertEducationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Education Details",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/education/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "education"
                ],
                "summary": "Update Education",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Education ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEducationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Education resp",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AuthReq": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.AuthResp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "dto.BaseResp": {
            "type": "object",
            "properties": {
                "error_schema": {
                    "$ref": "#/definitions/dto.ErrorSchema"
                },
                "output_schema": {
                    "description": "OutputSchema fleksibel (bisa berisi apa saja)"
                }
            }
        },
...
        "dto.ErrorMessage": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorSchema": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "$ref": "#/definitions/dto.ErrorMessage"
                }
            }
        },
        "dto.InsertCertificationReq": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "credential_id": {
                    "type": "string"
                },
                "expiration_date": {
                    "type": "string"
                },
                "issue_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.InsertEducationReq": {
            "type": "object",
            "properties": {
                "degree": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "field_of_study": {
                    "type": "string"
                },
                "school_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateCertificationReq": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "credential_id": {
                    "type": "string"
                },
                "expiration_date": {
                    "type": "string"
                },
                "issue_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateEducationReq": {
            "type": "object",
            "properties": {
                "degree": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "field_of_study": {
                    "type": "string"
                },
                "school_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserRegisterReq": {
            "type": "object",
            "properties": {
                "fullname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}